Reverting javascript changes to take a different approach
