+++ adminfunctions.php
@@ -40,39 +40,6 @@ function foxyshop_load_admin_scripts($hook) {
 	//Style - Always Do This
 	wp_enqueue_style('foxyshop_admin_css', FOXYSHOP_DIR . '/css/foxyshop-admin.css');
 
-	//adding admin tools
-	wp_enqueue_script('sorter', FOXYSHOP_DIR . '/js/tablesorter.js', array('jquery'));
-	$inlinesorter = 'var thenonce="'.wp_create_nonce('update-foxyshop-inventory').'";
-
-
-	jQuery(document).ready(function($){ '.foxyshop_manage_attributes_jquery('transaction').' });
-	';
-
-
-
-	 if (version_compare($foxyshop_settings['version'], '0.7.2', ">=") && $foxyshop_settings['domain']) { 
-$inlinesorter .= "
-
-jQuery(document).ready(function($){
-	$(\"#ajax_get_category_list\").click(function() {
-		var data = {
-			action: 'foxyshop_ajax_get_category_list',
-			security: '".wp_create_nonce("foxyshop-ajax-get-category-list")."'
-		};
-		$(\"#foxyshop_category_list_waiter\").show();
-		$.post(ajaxurl, data, function(response) {
-			if (response) {
-				$(\"#foxyshop_ship_categories\").val(response);
-			}
-			$(\"#foxyshop_category_list_waiter\").hide();
-		});
-
-	});
-
-});";
-	 } 
-
-
 	//Date Picker
 	if ($page == "foxyshop_order_management" || $page == "foxyshop_subscription_management") foxyshop_date_picker();
 
@@ -90,14 +57,6 @@ jQuery(document).ready(function($){
 		wp_enqueue_style('chosenStyle', FOXYSHOP_DIR . '/css/chosen.css');
 	}
 	foxyshop_date_picker();
-
-	wp_enqueue_script('foxyshop_tools', FOXYSHOP_DIR . '/js/all.js', array('jquery','sorter','jquery-ui-datepicker'));
-		wp_add_inline_script( 
-	'foxyshop_tools',
-	$inlinesorter,
-	'before' 
-);
-	wp_add_inline_script('foxyshop_tools','jQuery(document).ready(function($){ '. foxyshop_manage_attributes_jquery('subscription').'});','after');
 }
 
 //Loading in Public Style
@@ -124,7 +83,7 @@ function foxyshop_check_include_status() {
 		remove_action('init', 'foxyshop_insert_jquery');
 		if ($foxyshop_settings['include_exception_list'] != "*") {
 			add_action('wp_enqueue_scripts', 'foxyshop_remove_jquery', 99);
-			remove_action('wp_enqueue_scripts', 'foxyshop_insert_google_analytics', 100);
+			remove_action('wp_footer', 'foxyshop_insert_google_analytics', 100);
 		}
 	}
 }
@@ -145,38 +104,10 @@ function foxyshop_check_permalinks() {
 }
 
 
-function foxyshop_insert_allfrontend(){
-	global $foxyshop_settings;
-	if (!empty($foxyshop_settings['domain'])){
-	wp_enqueue_script('loaderforfoxy', "https://cdn.foxycart.com/" . esc_attr(str_replace('.foxycart.com','',$foxyshop_settings['domain'])) . "/loader.js",['jquery'],FOXYSHOP_VERSION);
-	}
-	wp_enqueue_script('foxyshop_public',FOXYSHOP_PLUGIN_URL.'js/all-public.js',['jquery'],FOXYSHOP_VERSION);
-	wp_enqueue_script('variationp',FOXYSHOP_PLUGIN_URL.'js/variation.process.js',['jquery'],FOXYSHOP_VERSION);
-	wp_enqueue_script('prettyphoto',FOXYSHOP_PLUGIN_URL.'js/prettyphoto/prettyPhoto.js',['jquery'],FOXYSHOP_VERSION);
-	wp_enqueue_style('prettyphoto',FOXYSHOP_PLUGIN_URL.'css/prettyphoto/prettyPhoto.css',[],FOXYSHOP_VERSION);
-	wp_enqueue_script('foxyshop_public_header',FOXYSHOP_PLUGIN_URL.'js/all-public-header.js',['jquery'],FOXYSHOP_VERSION);
-	wp_script_add_data( 'foxyshop_public_header', 'conditional', 'lt IE 9' );
-}
-
-
-function foxyshop_setidcheckout(){
-	$toadd = 'jQuery(document).ready(function($){
-	$("html").attr("id", "fc");
-});';
-	wp_add_inline_script( 
-	'foxyshop_public',
-	$toadd,
-	'after' 
-	);
-}
-
-
 
 //Insert Google Analytics
 function foxyshop_insert_google_analytics() {
 	global $foxyshop_settings;
-
-	$toadd = '';
 
 	//Advanced
 	if ($foxyshop_settings['ga_advanced']) {
@@ -186,15 +117,19 @@ function foxyshop_insert_google_analytics() {
 
 		//Universal
 		if ($foxyshop_settings['ga_type'] == "universal") {
-			$toadd = "(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+			?>
+<script>
+  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
   (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
   m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
   })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
 
-  ga('create', '".htmlspecialchars($foxyshop_settings['ga'])."', 'auto');";
-   if ($foxyshop_settings['ga_demographics']) $toadd .= "ga('require', 'displayfeatures');\n";
-  $toadd .= "ga('send', 'pageview');
- 
+  ga('create', '<?php echo htmlspecialchars($foxyshop_settings['ga']); ?>', 'auto');
+  <?php if ($foxyshop_settings['ga_demographics']) echo "ga('require', 'displayfeatures');\n" ?>
+  ga('send', 'pageview');
+
+</script>
+<script type="text/javascript" charset="utf-8">
   fcc.events.cart.preprocess.add(function(e, arr) {
     if (arr['cart'] == 'checkout' || arr['cart'] == 'updateinfo' || arr['output'] == 'json') {
       return true;
@@ -207,7 +142,7 @@ function foxyshop_insert_google_analytics() {
           });
         }
       });
-      return 'pause';
+      return "pause";
     }
     return true;
   });
@@ -217,23 +152,28 @@ function foxyshop_insert_google_analytics() {
         fcc.events.cart.process.resume();
       });
     });
-    return 'pause';
-  });";
+    return "pause";
+  });
+</script>
+<?php
+
+
 
 		//Legacy
 		} else {
 
-		$toadd = " 
+
+
+		?><script type="text/javascript" charset="utf-8">
 	var _gaq = _gaq || [];
-	_gaq.push(['_setAccount', '".htmlspecialchars($foxyshop_settings['ga'])."']);
-	_gaq.push(['_setDomainName', '".($_SERVER['SERVER_NAME'])."']);
-	_gaq.push(['_setAllowHash', 'false']);";
-	if (strpos($foxyshop_settings['domain'], '.foxycart.com') !== false) 
-		$toadd.= "_gaq.push(['_setAllowLinker', true]);\n"; 
-	$toadd .= "_gaq.push(['_trackPageview']);
+	_gaq.push(['_setAccount', '<?php echo htmlspecialchars($foxyshop_settings['ga']); ?>']);
+	_gaq.push(['_setDomainName', '<?php echo ($_SERVER['SERVER_NAME']); ?>']);
+	_gaq.push(['_setAllowHash', 'false']);
+	<?php if (strpos($foxyshop_settings['domain'], '.foxycart.com') !== false) echo "_gaq.push(['_setAllowLinker', true]);\n"; ?>
+	_gaq.push(['_trackPageview']);
 	(function() {
 		var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-		ga.src = ".apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';")."
+		ga.src = <?php echo apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';"); ?>
 		var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
 	})();
 	fcc.events.cart.preprocess.add(function(e, arr) {
@@ -252,7 +192,9 @@ function foxyshop_insert_google_analytics() {
 		jQuery.getJSON('https://' + storedomain + '/cart?' + fcc.session_get() + '&h:ga=' + escape(pageTracker._getLinkerUrl('', true)) + '&output=json&callback=?', function(data){});
 		return true;
 	});
-</script>";
+</script><?php
+
+
 		}
 
 
@@ -264,83 +206,90 @@ function foxyshop_insert_google_analytics() {
 
 		//Universal
 		if ($foxyshop_settings['ga_type'] == "universal") {
-		$toadd = "
+		?>
+
+<script>
   (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
   (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
   m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
   })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-  ga('create', '".foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']) )."', 'auto');
-  ".(($foxyshop_settings['ga_demographics']) ? "ga('require', 'displayfeatures');\n" : "")."
-  ga('send', 'pageview');";
+  ga('create', '<?php echo foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']) ) ; ?>', 'auto');
+  <?php if ($foxyshop_settings['ga_demographics']) echo "ga('require', 'displayfeatures');\n" ?>
+  ga('send', 'pageview');
+</script>
+
+
+		<?php
 		//Legacy
 		} else {
-		$toadd = "var _gaq = _gaq || [];
-_gaq.push(['_setAccount', '".foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']))."']);
+		?>
+
+		<script type="text/javascript">
+var _gaq = _gaq || [];
+_gaq.push(['_setAccount', '<?php echo foxy_wp_html(htmlspecialchars($foxyshop_settings['ga'])); ?>']);
 _gaq.push(['_trackPageview']);
 (function() {
 	var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-	ga.src = ".apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';")."
+	ga.src = <?php echo apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';"); ?>
 	var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-})();";
+})();
+</script><?php
 		}
 
 		} else {
-			$toadd = ("\n<!-- Google Analytics Not Loaded Because This is a Logged-In User -->\n");
+			echo ("\n<!-- Google Analytics Not Loaded Because This is a Logged-In User -->\n");
 		}
 	}
-	wp_add_inline_script( 
-	'foxyshop_public',
-	$toadd,
-	'after' 
-	);
 }
 
 //Google Analytics For Checkout
 function foxyshop_insert_google_analytics_checkout() {
 	global $foxyshop_settings;
 	if (!$foxyshop_settings['ga_advanced']) return;
-	$toadd = "";
-
 	if (version_compare($foxyshop_settings['version'], '2.0', ">=")) return;
 
 	//Universal
 	if ($foxyshop_settings['ga_type'] == "universal") {
 
-  $toadd .= "(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+		?>
+<script>
+  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
   (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
   m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
   })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
 
-  ga('create', '".foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']))."', 'auto', {
+  ga('create', '<?php echo foxy_wp_html(htmlspecialchars($foxyshop_settings['ga'])); ?>', 'auto', {
     'clientId': fc_json.custom_fields['ga'],
     'storage': 'none'
-  });";
-  if ($foxyshop_settings['ga_demographics']) $toadd .= "ga('require', 'displayfeatures');\n";
-  $toadd.= "ga('set', 'page', '/checkout');
+  });
+  <?php if ($foxyshop_settings['ga_demographics']) echo "ga('require', 'displayfeatures');\n" ?>
+  ga('set', 'page', '/checkout');
   ga('send', 'pageview');
- 
+
+</script>
+<script type="text/javascript" charset="utf-8">
   var originalValidateAndSubmit = FC.checkout.validateAndSubmit;
   function ga_tracker() {
     var wait = false;
-    if (typeof(fc_json.custom_fields['ga']) != \"undefined\") {
-      var method, method_co, source = \"\";
-      if (typeof(fc_json.custom_fields['_fcpm']) != \"undefined\" && fc_json.custom_fields['_fcpm'] != \"\") {
-        var fcpm = fc_json.custom_fields['_fcpm'].split(\"|\");
+    if (typeof(fc_json.custom_fields['ga']) != "undefined") {
+      var method, method_co, source = "";
+      if (typeof(fc_json.custom_fields['_fcpm']) != "undefined" && fc_json.custom_fields['_fcpm'] != "") {
+        var fcpm = fc_json.custom_fields['_fcpm'].split("|");
         method = fcpm[0];
         source = fcpm[1];
       }
-      if (jQuery(\"input[name='fc_payment_method']:first\").attr(\"type\") == \"radio\") {
-        method_co = jQuery(\"input[name='fc_payment_method']:checked\").val();
-      } else if (jQuery(\"input[name='fc_payment_method']:first\").attr(\"type\") == \"hidden\") {
-        method_co = jQuery(\"input[name='fc_payment_method']\").val();
+      if (jQuery("input[name='fc_payment_method']:first").attr("type") == "radio") {
+        method_co = jQuery("input[name='fc_payment_method']:checked").val();
+      } else if (jQuery("input[name='fc_payment_method']:first").attr("type") == "hidden") {
+        method_co = jQuery("input[name='fc_payment_method']").val();
       }
-      if (method != method_co || source == \"checkout\") {
+      if (method != method_co || source == "checkout") {
         method = method_co;
-        source = \"checkout\";
+        source = "checkout";
         wait = true;
-        jQuery.getJSON('https://' + document.domain + '/cart?' + FC.checkout.config.session + '&h:_fcpm=' + method + \"|\" + source + '&output=json&callback=?', function(cart) {
+        jQuery.getJSON('https://' + document.domain + '/cart?' + FC.checkout.config.session + '&h:_fcpm=' + method + "|" + source + '&output=json&callback=?', function(cart) {
           fc_json = cart;
-          if (method != \"plastic\" && method != \"purchase_order\" && source == \"checkout\") {
+          if (method != "plastic" && method != "purchase_order" && source == "checkout") {
             ga('send', 'pageview', '/' + method + '_payment');
           }
           originalValidateAndSubmit();
@@ -352,19 +301,31 @@ function foxyshop_insert_google_analytics_checkout() {
       originalValidateAndSubmit();
     }
   }
-  FC.checkout.override('validateAndSubmit', 'ga_tracker');";
+  FC.checkout.override('validateAndSubmit', 'ga_tracker');
+</script>
+
+
+
+		<?php
+
+
 
 	//Legacy
 	} else {
 
-	$toadd = "if (window.location.hash.search(/utma/) == -1 && typeof(fc_json.custom_fields['ga']) != \"undefined\") {
+	?>
+	<script type="text/javascript" charset="utf-8">
+		if (window.location.hash.search(/utma/) == -1 && typeof(fc_json.custom_fields['ga']) != "undefined") {
 			if (fc_json.custom_fields['ga'].length > 0) {
 				window.location.hash = fc_json.custom_fields['ga'].replace( /\&amp;/g, '&' );
 			}
-		} 
+		}
+	</script>
+
+	<script type="text/javascript">
 
 	  var _gaq = _gaq || [];
-	  _gaq.push(['_setAccount', '".foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']))."']);
+	  _gaq.push(['_setAccount', '<?php echo foxy_wp_html(htmlspecialchars($foxyshop_settings['ga'])); ?>']);
 	  _gaq.push(['_setDomainName', 'none']);
 	  _gaq.push(['_setAllowLinker', true]);
 	  _gaq.push(['_setAllowAnchor', true]);
@@ -372,65 +333,74 @@ function foxyshop_insert_google_analytics_checkout() {
 
 	  (function() {
 	    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-	    ga.src = ".apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';")."
+	    ga.src = <?php echo apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';"); ?>
 	    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
 	  })();
 
+	</script>
+
+	<script type="text/javascript" charset="utf-8">
 		function ga_tracker() {
-			if (typeof(fc_json.custom_fields['ga']) != \"undefined\" && jQuery('#fc_payment_method_paypal').attr(\"checked\") == true) {
+			if (typeof(fc_json.custom_fields['ga']) != "undefined" && jQuery('#fc_payment_method_paypal').attr("checked") == true) {
 				_gaq.push(['_trackPageview', '/paypal_payment']);
 				// setTimeout('return true;', 250); // TODO
 			}
 		}
-		FC.checkout.overload('validateAndSubmit', 'ga_tracker', null);";
+		FC.checkout.overload('validateAndSubmit', 'ga_tracker', null);
+	</script>
+	<?php
 	}
-	wp_add_inline_script( 
-	'foxyshop_public',
-	$toadd,
-	'after' 
-	);
 }
 
 
 //Google Analytics For Receipt
 function foxyshop_insert_google_analytics_receipt() {
 	global $foxyshop_settings;
-	$toadd = "";
 	if (!$foxyshop_settings['ga_advanced']) return;
 	if (version_compare($foxyshop_settings['version'], '2.0', ">=")) return;
 
 	//Universal
 	if ($foxyshop_settings['ga_type'] == "universal") {
 
-		$toadd="(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+		?>
+<script>
+  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
   (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
   m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
   })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
 
-  ga('create', '".foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']))."', 'auto', {
+  ga('create', '<?php echo foxy_wp_html(htmlspecialchars($foxyshop_settings['ga'])); ?>', 'auto', {
     'clientId': fc_json.custom_fields['ga'],
     'storage': 'none'
   });
-  var pageview = \"/receipt\";
-  if (typeof(fc_json.custom_fields['_fcpm']) != \"undefined\" && fc_json.custom_fields['_fcpm'] != \"\") {
-    var fcpm = fc_json.custom_fields['_fcpm'].split(\"|\");
-    if (fcpm[0] != \"plastic\" && fcpm[0] != \"purchase_order\") {
-      pageview += \"_\" + fcpm[0] + \"_from_\" + fcpm[1];
+  var pageview = "/receipt";
+  if (typeof(fc_json.custom_fields['_fcpm']) != "undefined" && fc_json.custom_fields['_fcpm'] != "") {
+    var fcpm = fc_json.custom_fields['_fcpm'].split("|");
+    if (fcpm[0] != "plastic" && fcpm[0] != "purchase_order") {
+      pageview += "_" + fcpm[0] + "_from_" + fcpm[1];
     }
   }
   ga('set', 'page', pageview);
   {% if first_receipt_display %}
-    ";
-    if($foxyshop_settings['ga_demographics']) $toadd.= "ga('require', 'displayfeatures');\n";
-  $toadd .= "ga('send', 'pageview');
-  {% endif %} 
+    <?php if ($foxyshop_settings['ga_demographics']) echo "ga('require', 'displayfeatures');\n" ?>
+    ga('send', 'pageview');
+  {% endif %}
+</script>
 {% if first_receipt_display %}
   ^^analytics_google_ga_universal^^
-{% endif %}";
+{% endif %}
+	<?php
+
+
+
 	//Legacy
 	} else {
-	$toadd = "var _gaq = _gaq || [];
-	  _gaq.push(['_setAccount', '".foxy_wp_html(htmlspecialchars($foxyshop_settings['ga']))."']);
+
+	?>
+	<script type="text/javascript">
+
+	  var _gaq = _gaq || [];
+	  _gaq.push(['_setAccount', '<?php echo foxy_wp_html(htmlspecialchars($foxyshop_settings['ga'])); ?>']);
 	  _gaq.push(['_setDomainName', 'none']);
 	  _gaq.push(['_setAllowLinker', true]);
 	  _gaq.push(['_setAllowAnchor', true]);
@@ -438,20 +408,17 @@ function foxyshop_insert_google_analytics_receipt() {
 
 	  (function() {
 	    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-	    ga.src = ".apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';")."
+	    ga.src = <?php echo apply_filters("ga_classic_src", "('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';"); ?>
 	    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
 	  })();
+
+	</script>
 
 	^^receipt_only_begin^^
 	^^analytics_google_ga_async^^
 	^^receipt_only_end^^
-	";
+	<?php
 	}
-	wp_add_inline_script( 
-	'foxyshop_public',
-	$toadd,
-	'after' 
-	);
 }
 
 //Product Category Comparison
@@ -1002,7 +969,8 @@ function foxyshop_manage_attributes($xml, $id, $att_type) {
 function foxyshop_manage_attributes_jquery($att_type) {
 	global $foxyshop_settings;
 	if (version_compare($foxyshop_settings['version'], '0.7.2', "<")) return "";
-	return `
+	?>
+
 	//Show New Form
 	$(".foxyshop_add_attribute").click(function(e) {
 		var id = $(this).attr("rel");
@@ -1097,6 +1065,8 @@ function foxyshop_manage_attributes_jquery($att_type) {
 
 		e.preventDefault();
 		return false;
-	});`;
+	});
+
+<?php
 
 }
