+++ categorysorting.php
@@ -128,5 +128,26 @@ function foxyshop_category_sort() {
 	}
 	?>
 	</div>
+<script type="text/javascript">
+function foxyshop_custom_order_load_event(){
+	jQuery("#foxyshop_category_order_list").sortable({
+		placeholder: "sortable-placeholder-category",
+		revert: false,
+		tolerance: "pointer",
+		update: function() {
+			var counter = 1;
+			jQuery("#foxyshop_category_order_list li").each(function() {
+				jQuery(this).find('.counter').html(counter);
+				counter++;
+			});
+		}
+	});
+};
+addLoadEvent(foxyshop_custom_order_load_event);
+function orderPages() {
+	jQuery("#updateText").html("<?php _e('Updating Category Order...', 'foxyshop') ?>");
+	jQuery("#foxyshop_category_order_value").val(jQuery("#foxyshop_category_order_list").sortable("toArray"));
+}
+</script>
 <?php
 }
