+++ foxyshop.php
@@ -93,8 +93,6 @@ if (is_admin()) {
 } else {
 	if ($foxyshop_settings['use_jquery']) add_action('wp_enqueue_scripts', 'foxyshop_insert_jquery', 15);
 
-	add_action('wp_enqueue_scripts', 'foxyshop_insert_allfrontend', 15);
-
 	if (version_compare($foxyshop_settings['version'], '2.0', ">=")) {
 		add_action('wp_footer', 'foxyshop_insert_foxycart_loader');
 	} else {
@@ -102,7 +100,7 @@ if (is_admin()) {
 	}
 	add_action('init', 'foxyshop_load_site_scripts', 1);
 	add_action('wp', 'foxyshop_check_include_status', 11);
-	if ($foxyshop_settings['ga']) add_action('wp_enqueue_scripts', 'foxyshop_insert_google_analytics', 100);
+	if ($foxyshop_settings['ga']) add_action('wp_footer', 'foxyshop_insert_google_analytics', 100);
 }
 
 //Setup Wizard
