+++ inventory.php
@@ -188,7 +188,61 @@ function foxyshop_inventory_management_page() {
 
 
 	</div>
- 
+
+<script type="text/javascript" src="<?php echo FOXYSHOP_DIR; ?>/js/tablesorter.js"></script>
+<script type="text/javascript">
+jQuery(document).ready(function($){
+	$(".inventory_update_width").blur(function() {
+		current_field_id = $(this).attr("id");
+		current_id = $("#" + current_field_id).attr("data-id");
+		new_count = $("#" + current_field_id).val();
+
+		$("#" + current_field_id).val(new_count);
+		$("#" + current_field_id).parents("tr").removeClass("inventory_update_width_highlight");
+
+		if (new_count != $("#original_count_" + current_id).val()) {
+
+			var data = {
+				action: 'save_inventory_values',
+				"_wpnonce": "<?php echo wp_create_nonce('update-foxyshop-inventory'); ?>",
+				"code": $("#code_" + current_id).val(),
+				"product_id": $("#productid_" + current_id).val(),
+				"new_count": new_count
+			};
+
+			$("#wait_" + current_id).addClass("waiting");
+			$.post(ajaxurl, data, function() {
+				$("#wait_" + current_id).removeClass("waiting");
+				$("#original_count_" + current_id).val(new_count);
+				$("#current_inventory_" + current_id).text(new_count);
+				if (new_count <= 0) {
+					$("#current_inventory_" + current_id).removeClass().addClass("inventoryU");
+				} else if (new_count <= parseInt($("#current_inventory_alert_" + current_id).text())) {
+					$("#current_inventory_" + current_id).removeClass().addClass("inventoryX");
+				} else {
+					$("#current_inventory_" + current_id).removeClass().addClass("inventoryA");
+				}
+			});
+		}
+  	});
+	$(".inventory_update_width").keypress(function(e) {
+		if (e.which == 13) {
+			$(this).trigger("blur");
+			return false;
+		}
+	});
+	$(".inventory_update_width").focus(function() {
+		$(this).parents("tr").addClass("inventory_update_width_highlight");
+	});
+	$("#inventory_level").tablesorter({
+		'cssDesc': 'asc sorted',
+		'cssAsc': 'desc sorted'
+	});
+
+
+});
+function foxyshop_format_number_single(num) { num = num.toString().replace(/\$|\,/g,''); if(isNaN(num)) num = "0"; sign = (num == (num = Math.abs(num))); num = Math.floor(num*100+0.50000000001); cents = num%100; num = Math.floor(num/100).toString(); if(cents<10) cents = "0" + cents; for (var i = 0; i < Math.floor((num.length-(1+i))/3); i++) num = num.substring(0,num.length-(4*i+3))+','+ num.substring(num.length-(4*i+3)); return (((sign)?'':'-') + num); }
+</script>
 
 
 <?php
