--- js/all-public-header.js
@@ -1,4 +1,0 @@
-var IEElms = ['article', 'aside', 'header', 'main', 'nav', 'section'];
-        for (var i = 0; i < IEElms.length; i++) {
-            document.createElement(IEElms[i]);
-        };
