+++ themefiles/foxyshop-single-category.php
@@ -54,6 +54,18 @@ This will allow you to upgrade FoxyShop without breaking your customizations. Mo
 	?>
 </div>
 <?php foxyshop_include('footer'); ?>
- 
+
+<script type="text/javascript">
+jQuery(document).ready(function($){
+	//Products
+	//This is set up for a two-column display. For a three column you need to do: nth-child(3n+1)
+	$(".foxyshop_product_list>li:nth-child(odd)").css("clear","left");
+
+	//Subcategories
+	//This is set up for a three-column display. For a two column you need to do: nth-child(odd)
+	$(".foxyshop_categories>li:nth-child(3n+1)").css("clear","left");
+
+});
+</script>
 
 <?php get_footer(); ?>
