+++ widgetcode.php
@@ -332,7 +332,14 @@ function foxyshop_dashboard_stats() {
 
 	echo '</div>'."\n";
 	?>
- 
+
+	<script type="text/javascript">
+	jQuery(document).ready(function($){
+		$.post(ajaxurl, { action: 'foxyshop_order_history_dashboard_action', security: '<?php echo wp_create_nonce("foxyshop-order-info-dashboard"); ?>' }, function(response) {
+			$("#foxyshop_dashboard_order_history").html(response)
+		});
+	});
+	</script>
 	<?php
 	echo '<div style="clear: both;"></div>'."\n";
 }
